// shared/types/common.ts
// SPDX-License-Identifier: MIT
// Copyright (c) 2026 CtxPilot

export interface OfficerStats {
  leadership: number;
  war: number;
  intelligence: number;
  politics: number;
  charisma: number;
}

export interface HexCoord {
  q: number;
  r: number;
}

export interface Point {
  x: number;
  y: number;
}